scaffold RetentionQueryContext
